Move GRC wizards to Governance folder
