+++ frontend/src/wizards/Governance/Policy/policyWizard.test.tsx
@@ -7,11 +7,11 @@ import {
   mockManagedClusters,
   mockPlacements,
   mockPolicy,
-} from '../../routes/Governance/governance.sharedMocks'
+} from '../../../routes/Governance/governance.sharedMocks'
 
 import { isExistingTemplateName, PolicyWizard } from './PolicyWizard'
 import { IResource } from '@patternfly-labs/react-form-wizard'
-import { PlacementKind } from '../../resources'
+import { PlacementKind } from '../../../resources'
 
 describe('ExistingTemplateName', () => {
   test('should return false for non-existing name', () => {
