+++ frontend/src/wizards/Governance/Policy/stable/CA-Security-Assessment-and-Authorization/policy-compliance-operator-install.yaml
