+++ frontend/src/wizards/Governance/Policy/stable/CM-Configuration-Management/policy-compliance-operator-cis-scan.yaml
