+++ frontend/src/wizards/Governance/Policy/stable/CM-Configuration-Management/policy-compliance-operator-e8-scan.yaml
