+++ frontend/src/wizards/Governance/Policy/stable/CM-Configuration-Management/policy-kyverno-add-network-policy.yaml
