+++ frontend/src/wizards/Governance/Policy/stable/CM-Configuration-Management/policy-namespace.yaml
