+++ frontend/src/wizards/Governance/Policy/stable/SC-System-and-Communications-Protection/policy-limitmemory.yaml
