+++ frontend/src/wizards/Governance/PolicyAutomation/PolicyAutomationWizard.test.tsx
@@ -4,7 +4,7 @@ import { RecoilRoot } from 'recoil'
 import { MemoryRouter } from 'react-router-dom'
 import { render, waitFor, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { clusterCuratorsState, subscriptionOperatorsState } from '../../atoms'
+import { clusterCuratorsState, subscriptionOperatorsState } from '../../../atoms'
 
 const mockGetwizardsynceditor = jest.fn()
 const mockCreatecredentialscallback = jest.fn()
